Reduce storage requirements for alf coeffs + variables renaming

+++ libovvc/nvcl_structures.h
@@ -737,12 +737,12 @@ typedef struct OVALFData
     uint8_t alf_cc_cr_filter_signal_flag;
 
     uint8_t alf_luma_coeff_delta_idx[MAX_NUM_ALF_CLASSES];
-    int16_t alf_luma_coeff[MAX_NUM_ALF_CLASSES][MAX_NUM_ALF_LUMA_COEFF];
-    int16_t alf_luma_clip_idx[MAX_NUM_ALF_CLASSES][MAX_NUM_ALF_LUMA_COEFF];
+    int8_t alf_luma_coeff[MAX_NUM_ALF_CLASSES][MAX_NUM_ALF_LUMA_COEFF];
+    uint8_t alf_luma_clip_idx[MAX_NUM_ALF_CLASSES][MAX_NUM_ALF_LUMA_COEFF];
     uint8_t alf_chroma_clip_flag;
     uint8_t alf_chroma_num_alt_filters_minus1;
-    int16_t alf_chroma_coeff[MAX_NUM_ALF_ALTERNATIVES_CHROMA][MAX_NUM_ALF_CHROMA_COEFF];
-    int16_t alf_chroma_clip_idx[MAX_NUM_ALF_ALTERNATIVES_CHROMA][MAX_NUM_ALF_CHROMA_COEFF];
+    int8_t alf_chroma_coeff[MAX_NUM_ALF_ALTERNATIVES_CHROMA][MAX_NUM_ALF_CHROMA_COEFF];
+    uint8_t alf_chroma_clip_idx[MAX_NUM_ALF_ALTERNATIVES_CHROMA][MAX_NUM_ALF_CHROMA_COEFF];
 
     uint8_t alf_cc_cb_filters_signalled_minus1;
     uint8_t alf_cc_cr_filters_signalled_minus1;
